task: SoC with controllers of books
